fixbug

+++ FTM.WebApi/Controllers/AuthenticateController.cs
@@ -2,8 +2,12 @@
 using FTM.WebApi.Models;
 using FTM.WebApi.Utility;
 using Google.Apis.Calendar.v3;
+using Google.Apis.Calendar.v3.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
+using System.Collections;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -16,12 +20,14 @@ namespace FTM.WebApi.Controllers
         private readonly ClientInfo clientInfo;
         private readonly FtmDbContext context;
         private readonly FtmDataStore dataStore;
+        private readonly IConfiguration configuration;
 
-        public AuthenticateController(ClientInfo clientInfo, FtmDbContext context, FtmDataStore dataStore)
+        public AuthenticateController(ClientInfo clientInfo, FtmDbContext context, FtmDataStore dataStore, IConfiguration configuration)
         {
             this.clientInfo = clientInfo;
             this.context = context;
             this.dataStore = dataStore;
+            this.configuration = configuration;
         }
 
         [HttpGet]
@@ -40,46 +46,43 @@ namespace FTM.WebApi.Controllers
                 {
                     if (result.Items.Any())
                     {
-                        foreach (var calendar in result.Items)
-                        {
-                            FtmCalendarInfo room = new FtmCalendarInfo()
-                            {
-                                CalendarId = calendar.Id,
-                                CalendarName = calendar.Summary,
-                                Description = calendar.Description,
-                                IsUseable = false
-                            };
-                            context.FtmCalendarInfo.Add(room);
-                        }
-                        await context.SaveChangesAsync();
+                        await Save(result.Items);
                     }
                 }
                 else
                 {
-                    if (calendars.Length != result.Items.Count)
+                    var calendarDb = calendars.Select(x => x.CalendarId);
+                    if (calendars.Length != result.Items.Count && result.Items.Any(x => !calendarDb.Contains(x.Id)))
                     {
                         //Remove all room
                         context.FtmCalendarInfo.RemoveRange(calendars);
 
-                        //Sync
-                        foreach (var calendar in result.Items)
-                        {
-                            FtmCalendarInfo room = new FtmCalendarInfo()
-                            {
-                                CalendarId = calendar.Id,
-                                CalendarName = calendar.Summary,
-                                Description = calendar.Description,
-                                IsUseable = false
-                            };
-                            context.FtmCalendarInfo.Add(room);
-                        }
-                        await context.SaveChangesAsync();
+                        await Save(result.Items);
                     }
                 }
 
                 transaction.Commit();
                 return Ok();
             }
+        }
+
+
+        private async Task Save(IEnumerable<CalendarListEntry> calendars)
+        {
+            foreach (var calendar in calendars)
+            {
+                if (calendar.Id == configuration["Settings:AdminEmail"])
+                    continue;
+                FtmCalendarInfo room = new FtmCalendarInfo()
+                {
+                    CalendarId = calendar.Id,
+                    CalendarName = calendar.Summary,
+                    Description = calendar.Description,
+                    IsUseable = false
+                };
+                context.FtmCalendarInfo.Add(room);
+            }
+            await context.SaveChangesAsync();
         }
     }
 }
